usando bigdecimal

+++ src/main/java/br/univesp/analisedados/dto/responses/PibCo2DadosDto.java
@@ -6,9 +6,9 @@ import br.univesp.analisedados.dto.TendenciaCentralDto;
 
 public record PibCo2DadosDto (Integer ano, TendenciaCentralDto pib, TendenciaCentralDto pibPerCap, TendenciaCentralDto co2) {
 	public PibCo2DadosDto(Integer ano, 
-			Double tendPib, BigDecimal pibMin,BigDecimal pibMax,
-			Double tendPibPerCap, BigDecimal pibPerCapMin,BigDecimal pibPerCapMax,
-			Double tendCo2, BigDecimal Co2Min,BigDecimal Co2Max){
+			BigDecimal tendPib, BigDecimal pibMin,BigDecimal pibMax,
+			BigDecimal tendPibPerCap, BigDecimal pibPerCapMin,BigDecimal pibPerCapMax,
+			BigDecimal tendCo2, BigDecimal Co2Min,BigDecimal Co2Max){
 		this(ano, 
 				new TendenciaCentralDto(tendPib, pibMin,pibMax),
 				new TendenciaCentralDto(tendPibPerCap, pibPerCapMin,pibPerCapMax),
